Fix warning, that apepard after I made asset counts signed 32bit in: 8e7d2d281ebe8118a5864d0cf634a38d6254cc1b

+++ tools/AssetDB/AssetDB.cpp
@@ -267,7 +267,7 @@ bool AssetDB::AddTestData(size_t numMods, const AssetTypeCountsArr& assetCounts)
 			if (assetCounts[x])
 			{
 				AssetType::Enum type = static_cast<AssetType::Enum>(x);
-				for (size_t j = 0; j < assetCounts[x]; j++)
+				for (int32_t j = 0; j < assetCounts[x]; j++)
 				{
 					core::string name = randomAssetName(assetNameLenMin, assetNameLenMax);
